adding pegination

+++ controllers/admin.js
@@ -2,33 +2,55 @@ const product = require("../models/product");
 const Product = require("../models/product");
 const mongodb = require("mongodb");
 const ObjectId = mongodb.ObjectId;
+const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file')
+
+
 
 exports.getAddProduct = (req, res, next) => {
-
-  if(!req.session.isLoggedIn){
-    return res.redirect('/login') ;
+  if (!req.session.isLoggedIn) {
+    return res.redirect("/login");
   }
   res.render("admin/edit-product", {
     pageTitle: "Add Product",
     path: "/admin/add-product",
     editing: false,
     // isAuthenticated : req.session.isLoggedIn
-   
   });
 };
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+
+  if (!image) {
+    return res.status(422).render("admin/product-edit", {
+      pageTitle: "add-product",
+      path: "admin/add-product",
+      editing: false,
+      hasError: true,
+
+      product: {
+        title: title,
+        // image : image ,
+        price: price,
+        description: description,
+      },
+      errorMessage: "attach file not valid",
+      validation,
+    });
+  }
+  // console.log(imageUrl);
+  const imageUrl = image.path;
 
   const product = new Product({
     title: title,
     imageUrl: imageUrl,
     price: price,
     description: description,
-    userId : req.user
+    userId: req.user,
   });
 
   // req.user
@@ -88,24 +110,47 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updateTitle = req.body.title;
   const updatePrice = req.body.price;
-  const updateImage = req.body.imageUrl;
+  const image = req.file;
   const updateDesc = req.body.description;
 
-  Product.findById(prodId).then(product => {
+  const errors = validationResult(req);
 
-    if(product.userId.toString() !== req.user._id.toString()){
-      return res.redirect('/')
-    }
-    product.title = updateTitle ;
-    product.price = updatePrice ;
-    product.description = updateDesc ;
-    product.imageUrl = updateImage ;
-    return product.save() ;
-  })
-    .then((result) => {
-      console.log("UPDATE PRODUCT");
-      res.redirect("/admin/products");
+  if (!errors.isEmpty()) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Edit Product",
+      path: "/admin/edit-product",
+      editing: true,
+      hasError: true,
+      product: {
+        title: updatedTitle,
+        price: updatedPrice,
+        description: updatedDesc,
+        _id: prodId,
+      },
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array(),
+    });
+  }
+
+  Product.findById(prodId)
+    .then((product) => {
+      if (product.userId.toString() !== req.user._id.toString()) {
+        return res.redirect("/");
+      }
+      product.title = updateTitle;
+      product.price = updatePrice;
+      product.description = updateDesc;
+      // product.imageUrl = updateImage ;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl)
+        product.imageUrl = image.path;
+      }
+      return product.save().then((result) => {
+        console.log("UPDATE PRODUCT");
+        res.redirect("/admin/products");
+      });
     })
+
     .catch((err) => console.log(err));
 
   // product.save();
@@ -113,20 +158,28 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteproduct = (req, res, next) => {
   const prodId = req.body.productId;
+  Product.findById(prodId).then(product =>{
+    if(!product){
+      return next(new Error('Product not found')) ;
+    }
+    fileHelper.deleteFile(product.imageUrl)
+    return Product.deleteOne({ _id: prodId, userId: req.user._id })
 
-  Product.deleteOne({_id : prodId , userId  : req.user._id})
-    .then(() => {
-      console.log("PRODUCT DESTROYED");
-      res.redirect("/admin/products");
-    })
-    .catch((err) => console.log(err));
+  })
+  .then(() => {
+    console.log("PRODUCT DESTROYED");
+    res.redirect("/admin/products");
+  })
+  .catch((err) => console.log(err));
+
+  
 };
 exports.getProducts = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.find({ userId  : req.user._id})
-  .populate('userId')
+  Product.find({ userId: req.user._id })
+    .populate("userId")
     .then((products) => {
-      console.log(products) ;
+      console.log(products);
       res.render("admin/products", {
         prods: products,
         pageTitle: "Admin Products",
